[SIL] Require explicit releasenone.
Functions which which consume their arguments will release them on any
branches where the arguments are unused.  Users of `@_effects(readnone)`
and `@_effects(readonly)` may not consider this when annotating their
functions.  Require that functions annotated thus are also annotated
@_effects(releasenone).

In order to support this, adjust SILFunctionBuilder to allow distinct
non-custom effects and to interpret such "conflicting" guarantees to
provide the strongest guarantee.  For example, annotating a function
both `@_effects(readnone)` and `@_effects(releasenone)` is equivalent to
only annotating it `@_effects(readnone)`.