Account for whether artificial subclasses are skipped on type caching
MetadataReader caches types only with the metadata address as the key.
However a type lookup can be requested skipping artificial subclasses or
not. This makes the cached results incorrect if two requests for the
same type, but skipping subclasses on one and not on the other, are made.
Fix this by adding a second dimension to the cache key.

rdar://101519300
(cherry picked from commit 9151305ccd4f88d61b91bdd50c395c50522c29ba)