[CanOSSALife] Respect barrier boundary edges.
When canonicalizing the lifetime of a lexical value, deinit barriers are
respected.  This is done by walking backwards from destroys and adding
encountered deinit barriers to liveness.

Previously, barrier edges did not result in any additions to liveness on
the theory that they would be rediscovered.  This is not always true (as
in the case of dead defs).

Here, each barrier edge different from the def block results in
additions to liveness.  Specifically, it results in the back of the
single predecessor being added to liveness.

rdar://115410893