Use clang to build Tablegen
Build-script tries to always cross-compile. As a result, the native
tablegen build tries to use /usr/bin/cc and /usr/bin/c++, which on
Amazon Linux 2 is too old to build tablegen. Instead, we can use the
clang/clang++ installed on the builders at /opt/swift/5.8.1/usr/bin,
which is inserted first in the path on the build image.

This route is sort of the least of all evils. We're already forcing the
rest of the build to use that clang, and it is capable of building for
the build machine, so might as well use it to build tablegen that then
emits pieces for the host build.