[swift-syntax] Reduce number of jobs that verify generated files
We were checking that the generated files matched the code in CodeGeneration in a number of jobs, which was still a relict from when `gyb_syntax_support` lived in the swift repo and was thus possible to break with a PR to the swift repo. Now, CodeGeneration lives exclusively in the swift-syntax repo and thus the checked in generated files can only be broken by a swift-syntax change.

Reduce the number of jobs that verify the generated files to:
- swift-syntax macOS PR testing (no need to run CodeGeneration on macOS and Linux)
- `buildbot_all_platforms,tools=RA,stdlib=RA`, which is used fro the macOS Pull Request status bot so we get signal if two PRs in swift-syntax raced and broke the checked in generated files

And while at it, also only verify the formatting of swift-syntax on macOS. There’s no need to verify it on all platforms.