[5.10][Runtime] Lazily copy ivar list in initObjCClass.
There's often nothing that needs to be fixed up in the ivar list. In that case, we can avoid copying it. This saves time and memory, and allows the class rodata to be in immutable memory.

rdar://116189946
(cherry picked from commit 641870a205f5b7788164037c1dd3e1b1e4ec4a56)