[Sema] Update diagnostic about public imports of non-resilient modules
The compiler reports public imports of non-resilient modules from a
resilient module. Make sure that when imports default to internal, which
also implies Swift 6, this is treated as an error and the fixit simply
deletes the `public` keyword.

We keep using the AccessLevelOnImport flag only to report these as
errors in Swift 5 for early adopters.