[Build] Don't create a separate target for the object file directory.
When building Swift sources, don't make a separate target for the object
file directory, because it may end up with the same name as would be
generated by another invocation of `handle_swift_sources`.

Instead, add the command to make the directory to the custom target that
actually does the build, which should have a unique name.

rdar://115278959