[CMake] allow custom options when building the compiler with debug info
Currently those are hardcoded to `-g`, but in some internal configurations
we would like to change them -- given the limited scope, for now
leverage a CMake cache variable instead of a full fledged build-script
flag.

There are other part of the build system that hardcode `-g` (e.g. in
`SwiftCompilerSources` and `AddSwiftStdlib.cmake`), but we are not
interested in those at the moment -- we will address those in the future
if need be.

Supports rdar://116371011