[cxx-interop] ClangDirectLookup should not require the C++ struct to be already imported
Doing a name lookup into a C++ struct currently causes all of its base structs to be eagerly imported into Swift. This causes degraded performance and triggers "circular reference" errors when a base struct refers to a derived struct (e.g. has a method that returns a pointer to a derived struct).

rdar://116426238